Removed prediction from DMM.
