+++ examples/stein_bnn.py
@@ -1,3 +1,6 @@
+# Copyright Contributors to the Pyro project.
+# SPDX-License-Identifier: Apache-2.0
+
 import argparse
 from collections import namedtuple
 from functools import partial
