+++ examples/stein_mixture_dmm.py
@@ -1,20 +1,23 @@
-"""  Based on "Structured Inference Networks for Nonlinear State Space Models" by Krishnan, Shalit and Sontag. (AAAI 2017)
+# Copyright Contributors to the Pyro project.
+# SPDX-License-Identifier: Apache-2.0
+
+"""  Based on "Structured Inference Networks for Nonlinear State Space Models"
+  by Krishnan, Shalit and Sontag. (AAAI 2017)
 """
 
-import matplotlib.pyplot as plt
-
 import jax
-from jax.experimental import stax
 import jax.numpy as jnp
 import jax.ops
+import matplotlib.pyplot as plt
+from jax.example_libraries import stax
 
 import numpyro
+import numpyro.distributions as dist
 from numpyro.contrib.callbacks import Progbar
 from numpyro.contrib.einstein import SteinVI
 from numpyro.contrib.einstein.kernels import RBFKernel
-import numpyro.distributions as dist
 from numpyro.examples.datasets import JSB_CHORALES, load_dataset
-from numpyro.infer import Trace_ELBO, log_likelihood
+from numpyro.infer import Trace_ELBO
 from numpyro.optim import Adam
 
 numpyro.set_platform("gpu")
@@ -46,8 +49,8 @@ def batch_fun(step):
 
 def _one_hot_chorales(seqs, num_nodes=88):
     return jnp.sum(jnp.array((seqs[..., None] == jnp.arange(num_nodes + 1))), axis=-2)[
-        ..., 1:
-    ]
+           ..., 1:
+           ]
 
 
 def Emitter(hidden_dim1, hidden_dim2, out_dim):
@@ -189,17 +192,17 @@ def GRU(hidden_dim, W_init=stax.glorot_normal()):
 
 
 def model(
-    seqs,
-    seqs_rev,
-    lengths,
-    *,
-    latent_dim=32,
-    emission_dim=100,
-    transition_dim=200,
-    data_dim=88,
-    gru_dim=150,
-    annealing_factor=1.0,
-    predict=False
+        seqs,
+        seqs_rev,
+        lengths,
+        *,
+        latent_dim=32,
+        emission_dim=100,
+        transition_dim=200,
+        data_dim=88,
+        gru_dim=150,
+        annealing_factor=1.0,
+        predict=False
 ):
     batch_size, max_seq_length, *_ = seqs.shape
 
@@ -231,8 +234,8 @@ def model(
             numpyro.sample(
                 "z_aux",
                 dist.Normal(z_loc, z_scale)
-                .mask(jnp.expand_dims(masks, axis=-1))
-                .to_event(2),
+                    .mask(jnp.expand_dims(masks, axis=-1))
+                    .to_event(2),
                 obs=z,
             )
 
@@ -243,24 +246,24 @@ def model(
         numpyro.sample(
             "obs_x",
             dist.Bernoulli(emission_probs)
-            .mask(jnp.expand_dims(masks, axis=-1))
-            .to_event(2),
+                .mask(jnp.expand_dims(masks, axis=-1))
+                .to_event(2),
             obs=oh_x,
         )
 
 
 def guide(
-    seqs,
-    seqs_rev,
-    lengths,
-    *,
-    latent_dim=32,
-    emission_dim=100,
-    transition_dim=200,
-    data_dim=88,
-    gru_dim=150,
-    annealing_factor=1.0,
-    predict=False
+        seqs,
+        seqs_rev,
+        lengths,
+        *,
+        latent_dim=32,
+        emission_dim=100,
+        transition_dim=200,
+        data_dim=88,
+        gru_dim=150,
+        annealing_factor=1.0,
+        predict=False
 ):
     batch_size, max_seq_length, *_ = seqs.shape
     seqs_rev = jnp.transpose(seqs_rev, axes=(1, 0, 2))
@@ -284,8 +287,8 @@ def guide(
             numpyro.sample(
                 "z",
                 dist.Normal(z_loc, z_scale)
-                .mask(jnp.expand_dims(masks, axis=-1))
-                .to_event(2),
+                    .mask(jnp.expand_dims(masks, axis=-1))
+                    .to_event(2),
             )
 
 
@@ -313,15 +316,6 @@ if __name__ == "__main__":
     lengths, seqs = get_batch()
 
     negative_elbo = svgd.evaluate(state, seqs, _reverse_padded(seqs, lengths), lengths)
-
-    prediction = svgd.predict(
-        state,
-        seqs,
-        _reverse_padded(seqs, lengths),
-        lengths,
-        num_samples=1,
-        predict=True["obs_x"]["value"],
-    )
 
     print("ELBO", -negative_elbo / lengths.sum())
     ll = svgd.log_likelihood(state, seqs, _reverse_padded(seqs, lengths), lengths)[
